web/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.3; python_version >= '3.8'
aiohttp-retry==2.8.3; python_version >= '3.7'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
asgiref==3.7.2; python_version >= '3.7'
attrs==23.2.0; python_version >= '3.7'
brotli==1.1.0
cachetools==5.3.3; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
distro==1.9.0; python_version >= '3.6'
django==4.0.4; python_version >= '3.8'
django-admin-rangefilter==0.12.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-better-admin-arrayfield==1.4.2
django-cors-headers==4.3.1; python_version >= '3.8'
django-filter==23.5; python_version >= '3.7'
django-formtools==2.5.1; python_version >= '3.8'
django-google-maps==0.13.0
django-otp==1.3.0; python_version >= '3.7'
django-phonenumber-field==7.3.0; python_version >= '3.8'
django-two-factor-auth[phonenumbers]==1.16.0
djangorestframework==3.14.0; python_version >= '3.6'
drf-spectacular==0.27.1; python_version >= '3.6'
frozenlist==1.4.1; python_version >= '3.8'
google-api-core[grpc]==2.17.1; python_version >= '3.7'
google-auth==2.28.1; python_version >= '3.7'
google-cloud-core==2.4.1; python_version >= '3.7'
google-cloud-texttospeech==2.16.2; python_version >= '3.7'
google-cloud-translate==3.15.2; python_version >= '3.7'
googleapis-common-protos==1.62.0; python_version >= '3.7'
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0; python_version >= '3.5'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.4; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
importlib-resources==6.1.2; python_version >= '3.8'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
markdown==3.5.2; python_version >= '3.8'
messagebird==2.1.0
multidict==6.0.5; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.9'
onesignal-sdk==2.0.0; python_version >= '3.6'
openai==1.13.3; python_full_version >= '3.7.1'
packaging==23.2; python_version >= '3.7'
phonenumbers==8.13.31
proto-plus==1.23.0; python_version >= '3.6'
protobuf==4.25.3; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==2.6.3; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pypng==0.20220715.0
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-liquid==1.12.1; python_version >= '3.7'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
qrcode==7.4.2; python_version >= '3.7'
referencing==0.33.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rpds-py==0.18.0; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
scipy==1.12.0; python_version >= '3.9'
sentry-sdk==1.40.6
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.4.4; python_version >= '3.5'
tqdm==4.66.2; python_version >= '3.7'
twilio==9.0.0; python_full_version >= '3.7.0'
typing-extensions==4.10.0; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.1; python_version >= '3.6'
uvicorn==0.27.1; python_version >= '3.8'
whitenoise==6.6.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
